+++ snitch/__init__.py
@@ -20,7 +20,7 @@ __all__ = [
     "get_notification_model",
 ]
 
-__version__ = "1.0.4"
+__version__ = "1.1.1"
 
 
 def autodiscover():